tweaking of rules + id'd a lot
These changes are paving the way for score calculation although it is gonna be a PAIN an absolute PAIN to hard code each addition. I am 10000000% sure there is a smarter way to do this, but i do not want to risk the time trying to find a way online that i can comprehend and apply. With a deadline this close I do not have time to fiddle with js as much as id like and because it is such a fickle language, I am gonna run with what i know works. also fuck JS for not accepting single digit IDs. i tested ID="1", ID="11" and it didnt work but as soon as i change it to "a1" it wants to work.